chore: moved test_filter_device_list to test_filter.py
For #147
